hotel-room-rental-service
This website is a hotel room rental service project and this project was prepared for school homework.

The config.php file is created for the website's database connection. Change it according to your own mysql port to connect to the database properly.

There are registration and login forms on this website.
After registering and logging in, it will take you to the page with rooms for rent and there you can see the rooms for rent, the rooms are taken from the database.
If you press anyone's purchase button, you will rent that room and it will appear on your profile.
Your information and the information and numbers of the rooms you rent appear on the My Account page.

There is a boolean column in the database containing users, this is 1 if you are an admin.
If you are an admin, an admin panel button will be added to the home page and you can access the admin panel page. You can edit hotel information, users, customers and hotel rooms in the admin panel.

There may be vulnerabilities in this project; additionally, unreal hotel names, addresses and hotel promotion descriptions were taken from chatGPT and user passwords in the database were not hidden.